requirements.txt
aiohttp==3.7.4.post0
asgiref==3.3.4
async-timeout==3.0.1
attrs==21.2.0
certifi==2020.12.5
chardet==4.0.0
Django==1.11.25
django-appconf==1.0.4
django-countries==7.2.1
django-ipware==3.0.2
django-user-agents==0.4.0
djangorestframework==3.6.4
geoip2==4.2.0
idna==2.10
jsonfield==3.1.0
maxminddb==2.0.3
memcache==0.5.1
multidict==5.1.0
pkg-resources==0.0.0
pytz==2021.1
requests==2.25.1
simplejson==3.17.2
sqlparse==0.4.1
typing-extensions==3.10.0.0
ua-parser==0.10.0
urllib3==1.26.5
user-agents==2.2.0
yarl==1.6.3
